+++ packages/contracts-rfq/test/FastBridgeV2.GasBench.Src.t.sol
@@ -104,11 +104,7 @@ contract FastBridgeV2GasBenchmarkSrcTest is FastBridgeV2SrcBaseTest {
     }
 
     function test_bridge_token_withExclusivity() public {
-        tokenParamsV2.quoteRelayer = relayerA;
-        tokenParamsV2.quoteExclusivitySeconds = int256(EXCLUSIVITY_PERIOD);
-        tokenParamsV2.quoteId = bytes("Created by Relayer A");
-        tokenTx.exclusivityRelayer = relayerA;
-        tokenTx.exclusivityEndTime = block.timestamp + EXCLUSIVITY_PERIOD;
+        setTokenTestExclusivityParams(relayerA, EXCLUSIVITY_PERIOD);
         bridge({caller: userA, msgValue: 0, params: tokenParams, paramsV2: tokenParamsV2});
         assertEq(fastBridge.bridgeStatuses(getTxId(tokenTx)), IFastBridgeV2.BridgeStatus.REQUESTED);
         assertEq(srcToken.balanceOf(userA), initialUserBalanceToken - tokenParams.originAmount);
@@ -189,11 +185,7 @@ contract FastBridgeV2GasBenchmarkSrcTest is FastBridgeV2SrcBaseTest {
     }
 
     function test_bridge_eth_withExclusivity() public {
-        ethParamsV2.quoteRelayer = relayerA;
-        ethParamsV2.quoteExclusivitySeconds = int256(EXCLUSIVITY_PERIOD);
-        ethParamsV2.quoteId = bytes("Created by Relayer A");
-        ethTx.exclusivityRelayer = relayerA;
-        ethTx.exclusivityEndTime = block.timestamp + EXCLUSIVITY_PERIOD;
+        setEthTestExclusivityParams(relayerA, EXCLUSIVITY_PERIOD);
         bridge({caller: userA, msgValue: ethParams.originAmount, params: ethParams, paramsV2: ethParamsV2});
         assertEq(fastBridge.bridgeStatuses(getTxId(ethTx)), IFastBridgeV2.BridgeStatus.REQUESTED);
         assertEq(userA.balance, initialUserBalanceEth - ethParams.originAmount);