+++ packages/contracts-rfq/test/FastBridgeV2.Src.Exclusivity.t.sol
@@ -8,17 +8,8 @@ contract FastBridgeV2SrcExclusivityTest is FastBridgeV2SrcTest {
     uint256 public constant EXCLUSIVITY_PERIOD = 60 seconds;
 
     function createFixturesV2() public virtual override {
-        tokenParamsV2.quoteRelayer = relayerA;
-        tokenParamsV2.quoteExclusivitySeconds = int256(EXCLUSIVITY_PERIOD);
-        tokenParamsV2.quoteId = bytes("Created by Relayer A");
-        ethParamsV2.quoteRelayer = relayerB;
-        ethParamsV2.quoteExclusivitySeconds = int256(EXCLUSIVITY_PERIOD);
-        ethParamsV2.quoteId = bytes("Created by Relayer B");
-
-        tokenTx.exclusivityRelayer = relayerA;
-        tokenTx.exclusivityEndTime = block.timestamp + EXCLUSIVITY_PERIOD;
-        ethTx.exclusivityRelayer = relayerB;
-        ethTx.exclusivityEndTime = block.timestamp + EXCLUSIVITY_PERIOD;
+        setTokenTestExclusivityParams(relayerA, EXCLUSIVITY_PERIOD);
+        setEthTestExclusivityParams(relayerB, EXCLUSIVITY_PERIOD);
     }
 
     function bridge(address caller, uint256 msgValue, IFastBridge.BridgeParams memory params) public virtual override {