+++ docs/bridge/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.8](https://github.com/synapsecns/sanguine/compare/@synapsecns/bridge-docs@0.3.7...@synapsecns/bridge-docs@0.3.8) (2024-10-07)
+
+**Note:** Version bump only for package @synapsecns/bridge-docs
+
+
+
+
+
 ## [0.3.7](https://github.com/synapsecns/sanguine/compare/@synapsecns/bridge-docs@0.3.6...@synapsecns/bridge-docs@0.3.7) (2024-10-05)
 
 **Note:** Version bump only for package @synapsecns/bridge-docs