+++ packages/rfq-indexer/api/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.6](https://github.com/synapsecns/sanguine/compare/@synapsecns/rfq-indexer-api@1.0.5...@synapsecns/rfq-indexer-api@1.0.6) (2024-10-07)
+
+**Note:** Version bump only for package @synapsecns/rfq-indexer-api
+
+
+
+
+
 ## [1.0.5](https://github.com/synapsecns/sanguine/compare/@synapsecns/rfq-indexer-api@1.0.4...@synapsecns/rfq-indexer-api@1.0.5) (2024-10-03)
 
 **Note:** Version bump only for package @synapsecns/rfq-indexer-api